Move ThemingScope to its own file
