+++ src/System.Windows.Forms/src/System.Windows.Forms.csproj
@@ -40,6 +40,7 @@
     <Compile Include="..\..\Common\src\SafeNativeMethods.cs" Link="System\Windows\Forms\SafeNativeMethods.cs" />
     <Compile Include="..\..\Common\src\ScreenDC.cs" Link="Common\ScreenDC.cs" />
     <Compile Include="..\..\Common\src\SpanHelpers.cs" Link="Common\SpanHelpers.cs" />
+    <Compile Include="..\..\Common\src\ThemingScope.cs" Link="Common\ThemingScope.cs" />
     <Compile Include="..\..\Common\src\UnsafeNativeMethods.cs" Link="System\Windows\Forms\UnsafeNativeMethods.cs" />
     <Compile Include="..\..\Common\src\Interop\**\*.cs">
       <Link>Interop\%(RecursiveDir)\%(Filename)%(Extension)</Link>
